# docs/mobile_v1_openapi.yaml
openapi: 3.0.3
info:
  title: FichaYa - Mobile API v1
  version: 1.25.0
  description: |
    API REST para la aplicacion Flutter de control de asistencia.
    Todos los endpoints (excepto auth) requieren header `Authorization: Bearer <token>`.
    Los modulos complementarios Feedback (`/api/v1/feedback`) y SKAP (`/api/skap`)
    usan el mismo Bearer JWT mobile, aunque no cuelgan del prefijo `/api/v1/mobile`.
    La campana de Alertas del home agrupa adelantos, pedidos de mercaderia y feedback
    para mostrar novedades o aprobaciones pendientes sin cambiar los flujos existentes.
  contact:
    name: Equipo Backend
servers:
  - url: http://localhost:5000/api/v1/mobile
    description: Local
  - url: https://control-asistencia-backend-8gle.onrender.com/api/v1/mobile
    description: Produccion

# ---------------------------------------------------------------------------
# Security
# ---------------------------------------------------------------------------
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    VacacionesMovimiento:
      type: object
      properties:
        id:
          type: integer
        tipo:
          type: string
          enum: [tomado, compensatorio, ajuste]
        dias:
          type: number
        fecha_desde:
          type: string
          format: date
          nullable: true
        fecha_hasta:
          type: string
          format: date
          nullable: true
        estado:
          type: string
          enum: [pendiente, aprobado, rechazado]
        observacion:
          type: string
          nullable: true
    VacacionesMovimientosResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        anio:
          type: integer
          example: 2026
        movimientos:
          type: array
          items:
            $ref: '#/components/schemas/VacacionesMovimiento'
...
    OkErrorResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        error:
          type: string

    # --- Adelanto ---
...
security:
  - BearerAuth: []

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /vacaciones/movimientos:
    get:
      tags: [Vacaciones]
      summary: Movimientos de vacaciones por anio
      description: Lista movimientos `tomado`, `compensatorio` y `ajuste` del empleado autenticado.
      parameters:
        - name: anio
          in: query
          schema:
            type: integer
            minimum: 2000
            maximum: 2100
            example: 2026
          description: Anio a consultar. Si se omite, usa el anio actual del servidor.
      responses:
        '200':
          description: Movimientos de vacaciones
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VacacionesMovimientosResponse'
        '400':
          description: Anio invalido o empleado sin datos validos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkErrorResponse'
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkErrorResponse'
...